Refactor: Split BunProcess.cpp into BunProcess.cpp and BunProcessFunctions.cpp
Split the large BunProcess.cpp file into two files for better code organization:
- BunProcess.cpp: Class definition, setup code, and function declarations
- BunProcessFunctions.cpp: All 67 function implementations (getters, setters, host functions)

Changes:
- Created BunProcessFunctions.cpp with all function implementations
- Moved 48 JSC_DEFINE_HOST_FUNCTION implementations
- Moved 10 JSC_DEFINE_CUSTOM_GETTER implementations
- Moved 9 JSC_DEFINE_CUSTOM_SETTER implementations
- Moved helper functions (toFileURI, loadSignalNumberMap, constructReportObjectComplete, etc.)
- Added JSC_DECLARE_* declarations for all functions in BunProcess.cpp
- Added all necessary extern "C" declarations and forward declarations to BunProcessFunctions.cpp

The refactoring improves code organization by separating function implementations
while maintaining all functionality. All process tests pass (100/104 tests, with
failures unrelated to this change).
